requirement.txt
pandas-ta
yahooquery